exclude operations from FeedManager
